+++ src/Generators/Tests/Laravel/FeatureClassFactory.php
@@ -19,8 +19,8 @@ class FeatureClassFactory extends ClassFactory
     /**
      * {@inheritdoc}
      */
-    protected function makeNamespace(ReflectionClass $reflectionClass): string
+    public function getTestSubNamespace(): string
     {
-        return parent::makeNamespace($reflectionClass).'\\Feature';
+        return '\\Feature';
     }
 }